Update vote status when getting all votes of a space
